deployments first pass
