--- google_cloud_automlops/deployments/enums.py
@@ -1,44 +1,0 @@
-# Copyright 2024 Google LLC. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Sets global enums."""
-
-# pylint: disable=C0103
-# pylint: disable=line-too-long
-
-from enum import Enum
-
-
-class Deployer(Enum):
-    """Enum representing the available options for orchestration management."""
-
-    CLOUDBUILD = 'cloud-build'
-    GITHUB_ACTIONS = 'github-actions'
-    # GITLAB_CI = 'gitlab-ci'   # roadmap item
-    # JENKINS = 'jenkins'   # roadmap item
-
-
-class CodeRepository(Enum):
-    """Enum representing the available options for source code repositories."""
-
-    BITBUCKET = 'bitbucket'   # roadmap item
-    CLOUD_SOURCE_REPOSITORIES = 'cloud-source-repositories'
-    GITHUB = 'github'
-    GITLAB = 'gitlab'   # roadmap item
-
-
-class ArtifactRepository(Enum):
-    """Enum representing the available options for artifact repositories."""
-
-    ARTIFACT_REGISTRY = 'artifact-registry'
