+++ google_cloud_automlops/deployments/templates/cloudbuild/cloudbuild.yaml.j2
