+++ google_cloud_automlops/deployments/templates/github_actions/github_actions.yaml.j2
