+++ google_cloud_automlops/deployments/templates/gitlab_ci/gitlab-ci.yml.j2
