+++ google_cloud_automlops/deployments/templates/gitlab_ci/readme.md
